requirements.txt
Flask==3.1.0
Pillow>=10.4.0
Flask-Limiter>=3.8.0
Flask-Session>=0.8.0
APScheduler>=3.10.4
werkzeug>=3.1.0
gunicorn>=21.0.0
